% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mGSZspeedup.R
\docType{data}
\name{mGSZspeedup}
\alias{mGSZspeedup}
\title{Speedup between mGSZ and MIGSAmGSZ.}
\format{A list with some timing bencharmks.}
\description{
A list with several benchmark timings used to show MIGSAmGSZ's speedup 
against mGSZ.
}

