% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGSAdataPackage.R
\docType{data}
\name{MIGSAdata}
\alias{MIGSAdata}
\title{MIGSAdata package}
\description{
The MIGSAdata package provides several data objects needed by MIGSA package 
to correctly generate its vignette, and follow it step by step.
}
\author{
Juan C. Rodriguez \email{jcrodriguez@bdmg.com.ar}, Cristobal Fresno 
\email{cfresno@bdmg.com.ar}, Andrea S. Llera \email{allera@leloir.org.ar} 
and Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
\references{
\enumerate{
\item Rodriguez, J. C., Gonzalez, G. A., Fresno, C., Llera, A. S., & 
Fernandez, E. A. (2016). Improving information retrieval in functional 
analysis. Computers in Biology and Medicine, 79, 10-20.
}
}
\keyword{MIGSA}
\keyword{data}
\keyword{objects}

