% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tcgaRNAseqData.R
\docType{data}
\name{tcgaRNAseqData}
\alias{tcgaRNAseqData}
\title{TCGA breast cancer RNAseq data.}
\format{A list with gene expression as data.frame and subtypes as vector.}
\description{
TCGA breast cancer RNAseq expresion matrix and each subjects subtypes.
Each dataset subjects were classified using the PAM50 algorithm. Only Basal 
and Luminal A subjects were kept.
Details in how this dataset was obtained can be find in MIGSA's 
gettingTcgaData vignette.
}

